requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
matplotlib